单本爬虫的DB表格式变更

+++ novel-crawl2/src/main/java/com/ideaflow/noveldownload/novel/model/Chapter.java
@@ -7,7 +7,6 @@ import lombok.Data;
 @Data
 @Builder
 public class Chapter {
-
     private Long id; // 主键id
     @Builder.Default
     private Long bookId = 0L; // 小说id
@@ -16,8 +15,8 @@ public class Chapter {
     private String content;
     private String cleanContent;
     @Builder.Default
-    private Long wordCount = 0L; // 字数
+    private Integer wordCount = 0; // 字数
     @Builder.Default
-    private Long order = 0L;
+    private Integer order = 0;
 
 }
